requirements.txt
# Detection (SAM2)
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0

# Web UI
gradio>=4.0.0
